Move license-related code to a separate module
