+++ src/utils/authenticate.ts
@@ -10,7 +10,6 @@ import type {
 import { env, Uri, window } from "vscode";
 
 import { assertIsError } from "./assert.ts";
-import { execLocalStack } from "./cli.ts";
 
 /**
  * Registers a {@link UriHandler} that waits for an authentication token from the browser,
@@ -110,21 +109,6 @@ export async function saveAuthToken(
 			.then(() => {
 				outputChannel.show(true);
 			});
-	}
-}
-
-const LICENSE_VALIDITY_MARKER = "license validity: valid";
-
-export async function checkIsLicenseValid(outputChannel: LogOutputChannel) {
-	try {
-		const licenseInfoResponse = await execLocalStack(["license", "info"], {
-			outputChannel,
-		});
-		return licenseInfoResponse.stdout.includes(LICENSE_VALIDITY_MARKER);
-	} catch (error) {
-		outputChannel.error(error instanceof Error ? error : String(error));
-
-		return false;
 	}
 }
